+++ packages/app-andrew/src/components/commons/NavigationDots/index.tsx
@@ -3,6 +3,7 @@ import { NavigationDots as NavigationDotsUI } from '@gtms/ui/NavigationDots'
 import { NavigationDotsFullView } from '../NavigationDotsFullView'
 import { INavigationDotsProps, baseUIQuery } from 'queries'
 import { loadMyGroups } from '@gtms/state-user'
+import { GroupAvatarNoImage } from 'enums'
 
 export const NavigationDots: FC = () => {
   const [state, setState] = useState<INavigationDotsProps>(
@@ -30,7 +31,7 @@ export const NavigationDots: FC = () => {
   }
 
   return (
-    <NavigationDotsUI groups={state.groups}>
+    <NavigationDotsUI groups={state.groups} noImage={GroupAvatarNoImage}>
       <NavigationDotsFullView />
     </NavigationDotsUI>
   )