+++ packages/state-post/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.7](https://github.com/gtms-org/gtms-frontend/compare/@gtms/state-post@0.8.6...@gtms/state-post@0.8.7) (2020-07-24)
+
+**Note:** Version bump only for package @gtms/state-post
+
+
+
+
+
 ## [0.8.6](https://github.com/gtms-org/gtms-frontend/compare/@gtms/state-post@0.8.5...@gtms/state-post@0.8.6) (2020-07-22)
 
 **Note:** Version bump only for package @gtms/state-post