+++ packages/state-user/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.9](https://github.com/gtms-org/gtms-frontend/compare/@gtms/state-user@0.17.8...@gtms/state-user@0.17.9) (2020-07-24)
+
+
+### Bug Fixes
+
+* GEOT-374: show more info on my groups page, allow user to se ([#765](https://github.com/gtms-org/gtms-frontend/issues/765)) ([93309d2](https://github.com/gtms-org/gtms-frontend/commit/93309d21f245df19c4274ecfd898adcfff9dc01e))
+
+
+
+
+
 ## [0.17.8](https://github.com/gtms-org/gtms-frontend/compare/@gtms/state-user@0.17.7...@gtms/state-user@0.17.8) (2020-07-22)